src/main/java/com/github/httpserver/exception/PreconditionFailedException.java
package com.github.httpserver.exception;

import com.github.httpserver.helper.HttpResponseBuilder;
import com.github.httpserver.protocol.HttpResponse;
import com.github.httpserver.protocol.HttpStatus;

public class PreconditionFailedException extends HttpException {

    private static final HttpStatus STATUS = HttpStatus.HTTP_PRECONDITION_FAILED;

    public PreconditionFailedException() {
        super("The resource doesn't match one of the request conditions");
    }

    @Override
    public HttpStatus getHttpStatus() {
        return STATUS;
    }

    @Override
    public HttpResponse toHttpResponse() {
        return new HttpResponseBuilder()
                .setStatus(STATUS)
                .appendBodyAsHTML("h1", STATUS.toString())
                .appendBodyAsHTML("p", getMessage())
                .appendContentLengthHeader()
                .build();
    }
}
